adde failing test to diaplay user email on review page
