+++ app/models/review.rb
@@ -1,5 +1,6 @@
 class Review < ApplicationRecord
   belongs_to :restaurant
+  belongs_to :user
   validates :rating, presence: true,
   numericality: {only_integer: true, greater_than_or_equal_to: 1, less_than_or_equal_to: 5}
 end
