+++ db/migrate/20180221144642_create_reviews.rb
@@ -4,6 +4,7 @@ class CreateReviews < ActiveRecord::Migration[5.1]
       t.integer :rating
       t.text :comment
       t.references :restaurant, foreign_key: true
+      t.references :user, foreign_key: true
 
       t.timestamps
     end
