+++ spec/features/add_review_spec.rb
@@ -6,6 +6,7 @@ feature 'Add a review' do
  end
 
   scenario 'Add a review for an existing restaurant' do
+    sign_up("user@test.com", "password", "password")
     add_restaurant('Meat and Egg', 'Poached eggs on a lovely steak')
     add_restaurant('Lovely Veggies', 'Home-grown stuff for veggie lovers')
     page.all(:link, 'Show')[1].click
@@ -15,6 +16,7 @@ feature 'Add a review' do
     expect(page).to have_content('2')
     expect(page).to have_content('Bad vegetables')
     expect(page).not_to have_content('Meat and Egg')
+    expect(page).to have_content("user@test.com")
   end
 
 feature 'no rating is added to the review' do
