+++ spec/rails_helper.rb
@@ -10,8 +10,8 @@ require 'rspec/rails'
 require 'capybara'
 require 'capybara/rails'
 require 'capybara/rspec'
-require_relative 'features/restaurant_helper'
-require_relative 'features/review_helper'
+require_relative 'features/helper'
+
 
 # Requires supporting ruby files with custom matchers and macros, etc, in
 # spec/support/ and its subdirectories. Files matching `spec/**/*_spec.rb` are
